changed inventory class name back
