+++ Project/MyGameLibrary/MyGameLibrary.csproj
@@ -54,7 +54,7 @@
     <Compile Include="NPC.cs" />
     <Compile Include="Player.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Inventory2.cs" />
+    <Compile Include="Inventory.cs" />
     <Compile Include="Vector2.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
